Initial ChatGPT generation.
"I  need a script that will accept the collection name and a target parent collection name as parameters. The purpose of the script is to "move" the collection to a new parent collection. The script will find the records of the collection and the parent collection.  If the collection's isDefaultCollection field is not NULL or equal to 1, output a warning. The script will update the collection's parent column to the id_local of the parent collection's record. It will also need to update the collection's genealogy field to a combination of the parent record's genealogy field and the second part of the collection's genealogy field's existing value."